[feature]파일 분리 및 화살표 함수 수정
