+++ src/app/visit-record/layout.tsx
@@ -1,5 +1,7 @@
 import Layout from '@components/layout';
 
-export default function VisitRecordLayout({ children }: { children: React.ReactNode }) {
+const VisitRecordLayout = ({ children }: { children: React.ReactNode }) => {
   return <Layout>{children}</Layout>;
-}
+};
+
+export default VisitRecordLayout;
